platform/tb-gateway/lora-bridge/reliability_manager/requirements.txt
paho-mqtt== 2.1.0
